% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JN_method.R
\name{allpossibleregression}
\alias{allpossibleregression}
\title{function to fit all possible regression, not recommended for large experiments}
\usage{
allpossibleregression(y, XME, XSE, maxterms)
}
\arguments{
\item{y}{a vector of response}

\item{XME}{a matrix of ctive Main Effects from stage 1 JN method}

\item{XSE}{a mtraix of second order terms}

\item{maxterms}{a numeric value indicating the max terms in the design}
}
\value{
a list of:
     \itemize{
       \item bestModel- the matrix of regressors for the model chosen from all possible regression
       \item bestSSE- the SSE of the chosen model
       \item minAICc- AICCc of the chosen model
       \item best_beta- the parameter estimates of the chosen model
     }
}
\description{
function to fit all possible regression, not recommended for large experiments
}
