% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_lambda_telescoping.R
\name{bootstrap_confidence_set}
\alias{bootstrap_confidence_set}
\title{function to compute the bootstrap confidence region for the location of the optimal response}
\usage{
bootstrap_confidence_set(
  beta,
  Sigma,
  B = 1000,
  me_index,
  qe_index,
  two_fi_index,
  C_dtype,
  nfactor,
  C,
  model_order = "non_standard",
  lower,
  upper,
  fx,
  minimize = TRUE,
  rng = c(),
  num_cores = NA
)
}
\arguments{
\item{beta}{current estimate of the regression parameters.}

\item{Sigma}{the variance of the regression parameters estimated.}

\item{B}{Number of bootstrap samples}

\item{me_index}{vector of index of Main Effects to be included in the design for model_order "non_standard"}

\item{qe_index}{vector index of Quadratic Effects to be included in the design for model_order "non_standard"}

\item{two_fi_index}{list of index of Two Factor Interactions to be included in the design for model_order "non_standard"}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{nfactor}{number of experimental factors}

\item{C}{list of candidate space for each experimental factor in the design}

\item{model_order}{a string indicating the order of the regression model. "linear","quadratic" (Main Effects+Quadratic Effects), "2FI" (Main Effects + TwoFactor Interactions),"full", "non_standard"- for a user defined model order order}

\item{lower}{a vector of lower limits of the experimental region for all factors (lower level constraint to the optimization)}

\item{upper}{a vector of upper limits of the experimental region for all factors (upper level constraint to the optimization)}

\item{fx}{function to compute regression prediction t(x)%*%b}

\item{minimize}{a boolean value indicating whether to minimize or maximize. Default value is minimize}

\item{rng}{seed for random number generation}

\item{num_cores}{number of cores for parallel processing. Default value is NA, in which case Sys.getenv("PBS_RESC_TOTAL_PROCS") is used to get the number of cores in a HPC. Otherwise, the user needs to provide a value. Use 1 for serial execution.}
}
\value{
returns a list of
   \itemize{
       \item confidence_set_x_star- the x_star values for each bootstrap sample of regression parameters
       \item confidence_set_beta- the bootstrap sample of regression parameters
       }
}
\description{
function to compute the bootstrap confidence region for the optimal x values. The function uses bootstrap samples to sample from the multivariate normal distribution of the regression parameters.The confidence region is the subspace of the experimental region that contains the optimal x values for each sample in the bootstrap. The confidence region can be the entire experimental region when there is high uncertainty in the regression estimate.The code uses parallel computing to compute the confidence region and currently works only on linux machines.
}
\examples{
\dontrun{
  bootstrap_confidence_set(beta,Sigma,B=1000,me_index,qe_index,two_fi_index,C_dtype,nfactor,C,model_order="non_standard", lower, upper, fx, rng)
}
}
