\name{CompoundRSO-package}
\alias{CompoundRSO-package}
\alias{CompoundRSO}
\docType{package}
\title{\packageTitle{CompoundRSO}}
\description{\packageDescription{CompoundRSO}}
\section{Package Content}{\packageIndices{CompoundRSO}}
\author{\packageAuthor{CompoundRSO}}
\section{Maintainer}{\packageMaintainer{CompoundRSO}}
\keyword{package}
