% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_lambda_telescoping.R
\name{construct_des_compound_serial}
\alias{construct_des_compound_serial}
\title{function for initiating multiple starts of the Adaptive-RSO designs using coordinate exchange}
\usage{
construct_des_compound_serial(
  x_matrix_start,
  x_start,
  W_mat,
  n,
  C_dtype,
  C,
  model_order,
  type = "augment",
  D_opt_design,
  I_opt_design,
  me_index = c(),
  qe_index = c(),
  two_fi_index = c(),
  n_starts = 100,
  w_candidate = seq(0.05, 0.95, 0.05),
  D_opt_type = "non_Bayesian",
  K = 0
)
}
\arguments{
\item{x_matrix_start}{the design matrix of the specified order including the intercept of the random starting}

\item{x_start}{the settings matrix of the random starting design}

\item{W_mat}{Moment matrix computed for a given lambda}

\item{n}{total number of runs in the design}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{C}{list of candidate space for each experimental factor in the design}

\item{model_order}{a string indicating the order of the regression model. "linear","quadratic" (Main Effects+Quadratic Effects), "2FI" (Main Effects + TwoFactor Interactions),"full", "non_standard"- for a user defined model order order}

\item{type}{a string indicating the type of design augmentation. "non_augment" or "augment" (augment an existing design)}

\item{D_opt_design}{the D-optimal design matrix}

\item{I_opt_design}{the I-optimal design matrix}

\item{me_index}{vector of index of Main Effects to be included in the design for model_order "non_standard"}

\item{qe_index}{vector index of Quadratic Effects to be included in the design for model_order "non_standard"}

\item{two_fi_index}{list of index of Two Factor Interactions to be included in the design for model_order "non_standard"}

\item{n_starts}{number of random starts, default at 100}

\item{w_candidate}{vector of weights to be used in the computation of compound optimal design}

\item{D_opt_type}{string indicating the type of D-optimal design Bayesian or regular D-optimal design. "Bayesian" or "non_Bayesian"}

\item{K}{a diagonal matrix specifying the prior variance for Bayesian D-optimal design. By default set to 0.001 for potential terms and 0 for primary terms.}
}
\value{
return a list of best design matrix for each w in the w_candidate vector
 @examples
  construct_des_compound_serial=function(x_matrix_start,x_start,W_mat,n,C_dtype,C,model_order, type= "augment",D_opt_design,I_opt_design,me_index = c(),qe_index = c(),two_fi_index = c(), n_starts=100,w_candidate=seq(0.05,0.95,0.05),D_opt_type="non_Bayesian",K=0)
}
\description{
function for initiating multiple starts of a compound D-I design using coordinate exchange.This is the serial version of the code, by default a parallel version is used. The D-I optimal design is the exploration-exploitation trade-off design
}
