% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_unif.R
\name{construct_des_compound_unif}
\alias{construct_des_compound_unif}
\title{function for initiating multiple starts of an compound uniform design using coordinate exchange.This is the parallel version and works only on linux machines.}
\usage{
construct_des_compound_unif(
  x_matrix_start,
  x_start,
  n,
  C_dtype,
  C,
  model_order,
  type = "augment",
  aug_type = "not_foldover",
  D_opt_design,
  D_opt_x,
  x_unifopt,
  unif_crit,
  n_starts = 500,
  D_opt_type = "Bayesian",
  w_candidate = seq(0.05, 0.95, 0.05),
  use_w = FALSE,
  use_cpp = FALSE,
  K = 0,
  me_index_daug,
  qe_index_daug,
  two_fi_index_daug,
  update_all_rows = FALSE,
  D_opt_tresh = 0.8,
  U_opt_tresh = 0.8,
  w_tresh = 0.02,
  rng = c(),
  num_cores = NA
)
}
\arguments{
\item{x_matrix_start}{the design matrix of the specified order including the intercept of the random starting}

\item{x_start}{the settings matrix of the random starting design}

\item{n}{total number of runs in the design}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{C}{list of candidate space for each experimental factor in the design}

\item{model_order}{a string indicating the order of the regression model. "linear","quadratic" (Main Effects+Quadratic Effects), "2FI" (Main Effects + TwoFactor Interactions),"full", "non_standard"- for a user defined model order order}

\item{type}{a string indicating the type of design augmentation. "non_augment" or "augment" (augment an existing deisgn)}

\item{aug_type}{a string indicating whether the design has constraint or not. "foldover" or "not_foldover"}

\item{D_opt_design}{the D-optimal design matrix}

\item{D_opt_x}{the settings matrix of the D-optimal design}

\item{x_unifopt}{the settings matrix of the Uniform-optimal design}

\item{unif_crit}{a string indicating the discrepancy measure for uniform criterion. "MD2" or "WD2"}

\item{n_starts}{number of random starts, default at 100}

\item{D_opt_type}{a string indicating whether the D-optimal design type is "Bayesian" or "Non_Bayesian"}

\item{w_candidate}{a vector of w values to compute the compound uniform design for}

\item{use_w}{a boolean value indicating whether w_candidate should be used or the bisection search method for finding the compound optimal designs}

\item{use_cpp}{a boolean value indicating whether to use the CPP version of coordinate exchange}

\item{K}{a diagonal matrix specifying the prior variance in Bayesian D-optimal design set to 0.001 for potential terms and 0 for primary terms}

\item{me_index_daug}{vector of index of Main Effects to be included in the D-optimal design for model_order "non_standard"}

\item{qe_index_daug}{vector index of Quadratic Effects to be included in the D-optimal design for model_order "non_standard"}

\item{two_fi_index_daug}{list of index of Two Factor Interactions to be included in the D-optimal design for model_order "non_standard"}

\item{update_all_rows}{a boolean value indicating whether all the rows should be updated or only the augmented rows should be updated}

\item{D_opt_tresh}{a numeric value providing the lower bound of the treshold in bisection search for the D-optimal criteria. Default value is 0.8}

\item{U_opt_tresh}{a numeric value providing the lower bound of the treshold in bisection search for the U-optimal criteria. Default value is 0.8}

\item{w_tresh}{the threshold for difference in weights w for the bisection search to converge. The default value is 0.02}

\item{rng}{seed for random number generation for the random design construction starts}

\item{num_cores}{number of cores for parallel processing. Default value is NA, in which case Sys.getenv("PBS_RESC_TOTAL_PROCS") is used to get the number of cores in a HPC. Otherwise, the user needs to provide a value. Use 1 for serial execution.}
}
\value{
a list of compound optimal designs for individual weights w
 @examples
  construct_des_compound_unif(x_matrix_start,x_start,n,C_dtype,C,model_order, type= "augment",aug_type="not_foldover",D_opt_design,D_opt_x,x_unifopt,unif_crit,n_starts=100,D_opt_type="Bayesian",w_candidate=seq(0.05,0.95,0.05),use_w=FALSE,use_cpp=FALSE,K=0,me_index_daug, qe_index_daug,two_fi_index_daug,update_all_rows=FALSE)
}
\description{
function for initiating multiple starts of an compound uniform design using coordinate exchange.This is the parallel version and works only on linux machines.
}
