% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_lambda_telescoping.R
\name{construct_des_compound_wcand}
\alias{construct_des_compound_wcand}
\title{function for initiating multiple starts of the Adaptive-RSO}
\usage{
construct_des_compound_wcand(
  x_matrix_start,
  x_start,
  W_mat,
  n,
  C_dtype,
  C,
  model_order,
  type = "augment",
  D_opt_design,
  D_opt_x,
  I_opt_design,
  me_index = c(),
  qe_index = c(),
  two_fi_index = c(),
  n_starts = 500,
  w_candidate = seq(0.05, 0.95, 0.05),
  D_opt_type = "non_Bayesian",
  rng = c(),
  K = 0,
  me_index_daug,
  qe_index_daug,
  two_fi_index_daug,
  use_cpp = TRUE,
  num_cores = NA
)
}
\arguments{
\item{x_matrix_start}{the design matrix of the specified order including the intercept of the random starting}

\item{x_start}{the settings matrix of the random starting design}

\item{W_mat}{Moment matrix computed for a given lambda}

\item{n}{total number of runs in the design}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{C}{list of candidate space for each experimental factor in the design}

\item{model_order}{a string indicating the order of the regression model. "linear","quadratic" (Main Effects+Quadratic Effects), "2FI" (Main Effects + TwoFactor Interactions),"full", "non_standard"- for a user defined model order order}

\item{type}{a string indicating the type of design augmentation. "non_augment" or "augment" (augment an existing design)}

\item{D_opt_design}{the D-optimal design matrix}

\item{D_opt_x}{the settings matrix of the D-optimal design}

\item{I_opt_design}{the I-optimal design matrix}

\item{me_index}{vector of index of Main Effects to be included in the I-optimal design for model_order "non_standard"}

\item{qe_index}{vector index of Quadratic Effects to be included in the I-optimal design for model_order "non_standard"}

\item{two_fi_index}{list of index of Two Factor Interactions to be included in the I-optimal design for model_order "non_standard"}

\item{n_starts}{number of random starts, default at 100}

\item{w_candidate}{vector of weights to be used in the computation of compound optimal design}

\item{D_opt_type}{Type of D-optimal design Bayesian or regular D-optimal design. "Bayesian" or "non_Bayesian"}

\item{rng}{random seed}

\item{K}{a diagonal matrix specifying the prior variance in Bayesian D-optimal design set to 0.001 for potential terms and 0 for primary terms}

\item{me_index_daug}{vector of index of Main Effects to be included in the Bayes-Doptimal design for model_order "non_standard"}

\item{qe_index_daug}{vector index of Quadratic Effects to be included in the Bayes-Doptimal design for model_order "non_standard"}

\item{two_fi_index_daug}{list of index of Two Factor Interactions to be included in the Bayes-Doptimal design for model_order "non_standard"}

\item{use_cpp}{a boolean value indicating whether to use the CPP version of coordinate exchange}

\item{num_cores}{number of cores for parallel processing. Default value is NA, in which case Sys.getenv("PBS_RESC_TOTAL_PROCS") is used to get the number of cores in a HPC. Otherwise, the user needs to provide a value. Use 1 for serial execution.}
}
\value{
return a list of best matrix for random starts for each w in the candidate list
  @examples
  construct_des_compound_wcand(x_matrix_start,x_start,W_mat,n,C_dtype,C, model_order, type= "augment",D_opt_design,D_opt_x,I_opt_design,me_index = c(),qe_index = c(),two_fi_index = c(), n_starts=100,w_candidate=seq(0.05,0.95,0.05),D_opt_type="non_Bayesian",K=0,me_index_daug, qe_index_daug,two_fi_index_daug,use_cpp=TRUE)
}
\description{
function for initiating multiple starts of Adaptove-RSO designs using coordinate exchange algorithm.This is the parallel version and works only on linux machines. The Adaptive-RSO designs are compound optimal designs between Bayesian D-optimal designs and I-optimal designs.
}
