% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_D_Optimality_telescoping.R
\name{construct_des_D_Bayes}
\alias{construct_des_D_Bayes}
\title{function for initiating the coordinate exchange for a given number of random starts}
\usage{
construct_des_D_Bayes(
  x_matrix_start,
  x_start,
  n,
  C_dtype,
  C,
  model_order,
  type = "augment",
  freeze_rows = 0,
  me_index = c(),
  qe_index = c(),
  two_fi_index = c(),
  cubic_index = c(),
  quatric_index = c(),
  n_starts = 100,
  K,
  design = "Bayesian",
  rng = c(),
  use_cpp = TRUE,
  num_cores = NA
)
}
\arguments{
\item{x_matrix_start}{the design matrix of the specified order including the intercept of the random starting}

\item{x_start}{the settings matrix of the random starting design}

\item{n}{total number of runs in the design}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{C}{list of candidate space for each experimental factor in the design}

\item{model_order}{a string indicating the order of the regression model. "linear","quadratic" (Main Effects+Quadratic Effects), "2FI" (Main Effects + TwoFactor Interactions),"full", "non_standard"- for a user defined model order order}

\item{type}{a string indicating the type of design augmentation. "non_augment" or "augment" (augment an existing design)}

\item{freeze_rows}{the rows of the matrix to freeze while augmenting, 0 when type is "non_augment"}

\item{me_index}{vector of index of Main Effects to be included in the design for model_order "non_standard"}

\item{qe_index}{vector index of Quadratic Effects to be included in the design for model_order "non_standard"}

\item{two_fi_index}{list of index of Two Factor Interactions to be included in the design for model_order "non_standard"}

\item{cubic_index}{vector of index of Cubic Effects to be included in the design for model_order "non_standard"}

\item{quatric_index}{vector of index of Quatric Effects to be included in the design for model_order "non_standard"}

\item{n_starts}{number of random starts, default at 100}

\item{K}{a diagonal matrix specifying the prior variance for Bayesian D-optimal design. By default set to 0.001 for potential terms and 0 for primary terms.}

\item{design}{a string indicating the type of design constructed. "Non-Bayesian" or "Bayesian" optimal design. The major difference is in how singularity of the matrix is evaluated. In a Bayesian optimal design X'X+K(prior variance matrix) is checked for singularity}

\item{rng}{seed for random number generation for the random design construction starts}

\item{use_cpp}{a boolean value indicating whether to use the CPP version of coordinate exchange}

\item{num_cores}{number of cores for parallel processing. Default value is NA, in which case Sys.getenv("PBS_RESC_TOTAL_PROCS") is used to get the number of cores in a HPC. Otherwise, the user needs to provide a value. Use 1 for serial execution.}
}
\value{
a list of
 \itemize{
       \item Bayes_dopt_design- the best Bayesian D-optimal design matrix of the specified order including the intercept from the random starts
       \item Bayes_dopt_x- a matrix with factor levels of the best Bayesian D-optimal design from the random starts
       \item Bayes_dopt- The value of best objective value of the Bayesian D-optimal design from the random starts
       }
 @examples
  construct_des_D_Bayes(x_matrix_start,x_start,n,C_dtype, C,model_order, type= "augment", freeze_rows=0,me_index = c(),qe_index = c(),two_fi_index = c(),cubic_index = c(),quatric_index = c(),n_starts=500,K,design = "Bayesian")
}
\description{
function for initiating the coordinate exchange for a given number of random starts and returns the best design. This is a parallel version of the code and works only on linux machines. For windows the makeCluster needs to be modified. By default the parallel version of the code will be run.
}
