% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_lambda_telescoping.R
\name{construct_des_I_lambda}
\alias{construct_des_I_lambda}
\title{function for initiating the coordinate exchange for I optimal design for a given number of random starts}
\usage{
construct_des_I_lambda(
  x_matrix_start,
  x_start,
  W_mat,
  n,
  C_dtype,
  C,
  model_order,
  type = "augment",
  me_index = c(),
  qe_index = c(),
  two_fi_index = c(),
  n_starts = 100,
  rng = c(),
  use_cpp = TRUE,
  num_cores = NA
)
}
\arguments{
\item{x_matrix_start}{the design matrix of the specified order including the intercept of the random starting}

\item{x_start}{the settings matrix of the random starting design}

\item{W_mat}{Moment matrix computed for a given lambda}

\item{n}{total number of runs in the design}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{C}{list of candidate space for each experimental factor in the design}

\item{model_order}{a string indicating the order of the regression model. "linear","quadratic" (Main Effects+Quadratic Effects), "2FI" (Main Effects + TwoFactor Interactions),"full", "non_standard"- for a user defined model order order}

\item{type}{a string indicating the type of design augmentation. "non_augment" or "augment" (augment an existing design)}

\item{me_index}{vector of index of Main Effects to be included in the design for model_order "non_standard"}

\item{qe_index}{vector index of Quadratic Effects to be included in the design for model_order "non_standard"}

\item{two_fi_index}{list of index of Two Factor Interactions to be included in the design for model_order "non_standard"}

\item{n_starts}{number of random starts, default at 100}

\item{rng}{seed for random number generation}

\item{use_cpp}{a boolean value indicating whether to use the CPP version of coordinate exchange}

\item{num_cores}{number of cores for parallel processing. Default value is NA, in which case Sys.getenv("PBS_RESC_TOTAL_PROCS") is used to get the number of cores in a HPC. Otherwise, the user needs to provide a value. Use 1 for serial execution.}
}
\value{
a list of
 \itemize{
       \item best_I_opt_design- the best I-optimal design matrix of the specified order from the random starts
       \item best_Iopt_x- a matrix of factor levels for the best I-optimal design amongst the random starts
       \item best_I_obj- objective value of the best I-optimal design from the random starts
       }
 @examples
  construct_des_I_lambda(x_matrix_start,x_start,W_mat,n,C_dtype, C, model_order, type= "augment", me_index = c(),qe_index = c(),two_fi_index = c(), n_starts=100)
}
\description{
function for initiating the coordinate exchange for I optimal design a given number of random starts and returns the best design. This is the parallel version of the code and works only on linux machines.
}
