% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coor_exch_compound_unif}
\alias{coor_exch_compound_unif}
\title{function for computing the compound uniform design using coordinate exchange}
\usage{
coor_exch_compound_unif(
  x,
  x_matrix,
  C,
  model_order,
  C_dtype,
  freeze_rows,
  x_matrix_Dopt,
  x_unifopt,
  unif_crit,
  w,
  K,
  me_index_daug,
  qe_index_daug,
  two_fi_index_daug,
  telescoping = FALSE
)
}
\arguments{
\item{x}{settings matrix of the design}

\item{x_matrix}{a starting design matrix of the specified order including the intercept}

\item{C}{list of candidate points for all factors in the design}

\item{model_order}{string indicating linear,quadratic (Main Effects+Quadratic Effects), 2FI (Main Effects + TwoFactor Interactions),full, non_standard for a user defined order}

\item{C_dtype}{data type of the factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{freeze_rows}{the rows of the matrix to freeze while augmenting}

\item{x_matrix_Dopt}{the D-optimal design matrix}

\item{x_unifopt}{the settings matrix of the Uniform-optimal design}

\item{unif_crit}{uniform criterion one of "MD2" (default) or "WD2"}

\item{w}{weight for computing the compound optimal design}

\item{K}{a diagonal matrix of prior variance for the Bayesian D optimal design}

\item{me_index_daug}{vector of index of Main Effects to be included in the D-optimal design for model_order "non_standard"}

\item{qe_index_daug}{vector index of Quadratic Effects to be included in the D-optimal design for model_order "non_standard"}

\item{two_fi_index_daug}{list of index of Two Factor Interactions to be included in the D-optimal design for model_order "non_standard"}

\item{telescoping}{a boolean value indicating whether telescoping (fine tuning) the current design or not}
}
\value{
a list of x_matrix,prev,x_best,deff_best,Unif_best
  \itemize{
        \item x_matrix- the design matrix for a given model order
        \item opt_val- the compound optimal objective value
        \item x_best- matrix of factor levels for the optimal design
        \item deff_best- the d-efficiency of the optimal design
        \item Unif_best- the uniform efficiency of the optimal design
        }
}
\description{
function for computing the compound uniform design using coordinate exchange
}
\examples{
\dontrun{
coor_exch_compound_unif(x,x_matrix,C,model_order,C_dtype,freeze_rows=0,x_matrix_Dopt,x_unifopt,unif_crit="MD2",w,K=0,me_index_daug, qe_index_daug,two_fi_index_daug,telescoping=FALSE)
}
}
