% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_unif.R
\name{coor_exch_compound_unif_foldover}
\alias{coor_exch_compound_unif_foldover}
\title{function for computing the compound uniform design using coordinate exchange with foldover constraint}
\usage{
coor_exch_compound_unif_foldover(
  x,
  x_matrix,
  C,
  model_order,
  C_dtype,
  freeze_rows = 0,
  x_matrix_Dopt,
  x_unifopt,
  unif_crit = "MD2",
  w,
  K = 0,
  me_index_daug,
  qe_index_daug,
  two_fi_index_daug,
  telescoping = FALSE
)
}
\arguments{
\item{x}{settings matrix}

\item{x_matrix}{design matrix of the specified order including the intercept}

\item{C}{list of candidate space for each experimental factor in the design}

\item{model_order}{a string indicating the order of the regression model. "linear","quadratic" (Main Effects+Quadratic Effects), "2FI" (Main Effects + TwoFactor Interactions),"full", "non_standard"- for a user defined model order order}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{freeze_rows}{the rows of the matrix to freeze while augmenting, 0 when type is "non_augment"}

\item{x_matrix_Dopt}{the D-optimal design matrix}

\item{x_unifopt}{the Uniform-optimal design matrix}

\item{unif_crit}{string indicating the uniform optimal criterion. One of "MD2" (default) or "WD2"}

\item{w}{weight for computing the compound optimal design}

\item{K}{a diagonal matrix of prior variance for the Bayesian D optimal design}

\item{me_index_daug}{vector of index of Main Effects to be included in the D-optimal design for model_order "non_standard"}

\item{qe_index_daug}{vector index of Quadratic Effects to be included in the D-optimal design for model_order "non_standard"}

\item{two_fi_index_daug}{list of index of Two Factor Interactions to be included in the D-optimal design for model_order "non_standard"}

\item{telescoping}{a boolean value taking TRUE or FALSE. Telescoping reduces the candidate factors the local region of the optimum design}
}
\value{
a list of
 \itemize{
       \item Compound_unif_design- Comound Uniform optimal design matrix of the specified order including the intercept
       \item Compound_unif_obj- The value of the Comound Uniform Optimal objective value
       \item Compound_unif_x- a matrix with factor levels of the Comound Uniform optimal design
       \item Compound_Deff- Bayesian D effciency of the compound uniform design
       \item Compound_Ueff- Uniform Effciency of the compound uniform design
       }
 @examples
 \dontrun{
  coor_exch_compound_unif_foldover(x,x_matrix,C,model_order,C_dtype,freeze_rows=0,x_matrix_Dopt,x_unifopt,unif_crit="MD2",w,K=0,me_index_daug, qe_index_daug,two_fi_index_daug,telescoping=FALSE)
 }
 @export
}
\description{
function for computing the compound uniform design using coordinate exchange with foldover constraint
}
