% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coor_exch_D_Bayes}
\alias{coor_exch_D_Bayes}
\title{function for constructing a Bayesian D-optimal design using coordinate exchange}
\usage{
coor_exch_D_Bayes(
  x,
  x_matrix,
  C,
  model_order,
  C_dtype,
  freeze_rows,
  K,
  me_index,
  qe_index,
  two_fi_index,
  type = "non_augment",
  design = "Bayesian",
  telescoping = FALSE
)
}
\arguments{
\item{x}{settings matrix for the design}

\item{x_matrix}{design matrix of the specified order}

\item{C}{list of candidate space for each experimental factor in the design}

\item{model_order}{a string indicating the order of the regression model. "linear","quadratic" (Main Effects+Quadratic Effects), "2FI" (Main Effects + TwoFactor Interactions),"full", "non_standard"- for a user defined model order order}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{freeze_rows}{the rows of the matrix to freeze while augmenting, 0 when type is "non_augment"}

\item{K}{a diagonal matrix specifying the prior variance for Bayesian D-optimal design. By default set to 0.001 for potential terms and 0 for primary terms.}

\item{me_index}{vector of index of Main Effects to be included in the design for model_order "non_standard"}

\item{qe_index}{vector of index of Quadratic Effects to be included in the design for model_order "non_standard"}

\item{two_fi_index}{list of index of Two Factor Interactions to be included in the design for model_order "non_standard"}

\item{type}{a string indicating the type of design augmentation. "non_augment" or "augment" (augment an existing deisgn)}

\item{design}{a string indicating the type of design constructed. "Non-Bayesian" or "Bayesian" optimal design. The major difference is in how signularity of the matrix is evaluated. In a Bayesian optimal design X'X+K(prior variance matrix) is checked for singularity}

\item{telescoping}{a boolean value taking TRUE or FALSE. Telescoping reduces the candidate factors the local region of the optimum design}
}
\value{
a list of
 \itemize{
       \item x_matrix- a Bayesian D-optimal design matrix of the specified order
       \item optimal_value- Optimal value of the objective value of the Bayesian D-optimal design
       \item x- a matrix with factor levels of the Bayesian D-optimal design
       }
 @examples
 \dontrun{
  coor_exch_D_Bayes(x,x_matrix,C,model_order,C_dtype,freeze_rows = freeze_rows,me_index=c(),qe_index=c(),two_fi_index=c(),cubic_index = c(),quatric_index = c(),type="non_augment",telescoping=FALSE,K)
 }
 @export
}
\description{
function for constructing a Bayesian D-optimal design using coordinate exchange
}
