% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coor_exch_I_augment}
\alias{coor_exch_I_augment}
\title{function for constructing an I-optimal design using coordinate exchange}
\usage{
coor_exch_I_augment(
  x,
  x_matrix,
  C,
  model_order,
  C_dtype,
  W,
  freeze_rows,
  me_index,
  qe_index,
  two_fi_index,
  type = "augment",
  telescoping = FALSE
)
}
\arguments{
\item{x}{settings matrix of the design}

\item{x_matrix}{design matrix of the specified order including the intercept}

\item{C}{list of candidate space for each experimental factor in the design}

\item{model_order}{a string indicating the order of the regression model. "linear","quadratic" (Main Effects+Quadratic Effects), "2FI" (Main Effects + TwoFactor Interactions),"full", "non_standard"- for a user defined model order order}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{W}{Moment matrix computed for a given lambda}

\item{freeze_rows}{the rows of the matrix to freeze while augmenting, 0 when type is "non_augment"}

\item{me_index}{vector of index of Main Effects to be included in the design for model_order "non_standard"}

\item{qe_index}{vector index of Quadratic Effects to be included in the design for model_order "non_standard"}

\item{two_fi_index}{list of index of Two Factor Interactions to be included in the design for model_order "non_standard"}

\item{type}{a string indicating the type of design augmentation. "non_augment" or "augment" (augment an existing deisgn)}

\item{telescoping}{a boolean value taking TRUE or FALSE. Telescoping reduces the candidate factors the local region of the optimum design}
}
\value{
a list of
 \itemize{
       \item Iopt_design- I-optimal design matrix of the specified order including the intercept from the random starts
       \item Iopt_obj- The value of the I-optimal objective value
       \item Iopt_x- a matrix with factor levels of the I-optimal design
       }
 @examples
  coor_exch_I_augment(x,x_matrix,C,model_order,C_dtype, W, freeze_rows=0,type="non_augment",me_index=c(),qe_index=c(),two_fi_index=c(),telescoping=FALSE)
}
\description{
function for constructing an I-optimal design using coordinate exchange
}
