% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JN_method.R
\name{create_interaction_matrix}
\alias{create_interaction_matrix}
\title{an internal function to create interaction matrix for model selection}
\usage{
create_interaction_matrix(
  C_active_dtype,
  C_active_names,
  C_non_active_names,
  x_matrix_me
)
}
\arguments{
\item{C_active_dtype}{a string vector containing the datatype of the active ME}

\item{C_active_names}{a string vector names of active experimental factor}

\item{C_non_active_names}{a string vector names of inactive experimental factors. It is not NA only when a weak heredity interaction matrix is desired.}

\item{x_matrix_me}{a matrix of ME for the design}
}
\value{
a matrix of interaction terms
}
\description{
constructs the interaction matrix for model selection
}
