% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_X_matrix_v2.0.R
\name{delta_D}
\alias{delta_D}
\title{an internal function to compute the change in determinant value for a coordinate exhange in a D-optimal design}
\usage{
delta_D(x, x_i, D)
}
\arguments{
\item{x}{the row vector with the new coordinate value}

\item{x_i}{the current row vector that needs to be exchanged}

\item{D}{current X'X inverse values}
}
\description{
an internal function to compute the change in determinant value for a coordinate exhange in a D-optimal design
}
