% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_lambda_telescoping.R
\name{delta_I}
\alias{delta_I}
\title{internal function to compute the change in the objective function of I-optimal design for each coordinate exchange}
\usage{
delta_I(x, x_i, D, W)
}
\arguments{
\item{x}{the row vector with the new coordinate value}

\item{x_i}{the current row vector that needs to be exchanged}

\item{D}{Current (X'X)inv matrix}

\item{W}{Moment matrix computed for a given lambda}
}
\value{
returns the the change in I optimal objective function for the coordinate exchange
}
\description{
internal function to compute the change in the objective function of I-optimal design for each coordinate exchange
}
