% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_unif.R
\name{discW2_update}
\alias{discW2_update}
\title{computes the W2 uniform discrepancy criteria update for each coordinate exchange}
\usage{
discW2_update(X, i, j, xij_new, disc_matrix, update = FALSE)
}
\arguments{
\item{X}{settings matrix}

\item{i}{row being updated}

\item{j}{column being updated}

\item{xij_new}{new factor level at the i,j index}

\item{disc_matrix}{discrepancy matrix for W2 discrepancy measure}

\item{update}{boolean value indicating whether to update or create a new discrepancy matrix}
}
\value{
returns a discrepancy matrix
}
\description{
computes the W2 uniform discrepancy criteria update for each coordinate exchange
}
\examples{
\dontrun{
discW2_update(X,i,j,xij_new,disc_matrix,update=FALSE)
}
}
