% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{discW2_update_cpp}
\alias{discW2_update_cpp}
\title{computes the W2 uniform discrepancy criteria update for each coordinate exchange}
\usage{
discW2_update_cpp(
  X,
  update,
  disc_matrix,
  rescale = TRUE,
  i = 0L,
  j = 0L,
  xij_new = 0L
)
}
\arguments{
\item{X}{settings matrix of the design}

\item{update}{boolean value indicating whether to update or create a new discrepancy matrix}

\item{disc_matrix}{discrepancy matrix for W2 discrepancy measure}

\item{rescale}{boolean value indicating whether to rescale the design matrix to a (0,1) experimental region}

\item{i}{row being updated}

\item{j}{column being updated}

\item{xij_new}{new factor level at the i,j index}
}
\value{
returns a discrepancy matrix
}
\description{
computes the W2 uniform discrepancy criteria update for each coordinate exchange
}
\examples{
\dontrun{
discW2_update(X,i,j,xij_new,disc_matrix,update=FALSE)
}
}
