% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sumulation_Functions.R
\name{dsd_model_selection}
\alias{dsd_model_selection}
\title{selects the active first-order and second-order terms for the stage 1 design using JN model selection method}
\usage{
dsd_model_selection(
  design_matrix,
  fake_factor_index,
  C_dtype,
  heredity = "strong",
  model_selection_type = "JN_org",
  rm_valruns = c(),
  y
)
}
\arguments{
\item{design_matrix}{DSD settings matrix with fake factors}

\item{fake_factor_index}{index of fake factor columns}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{heredity}{string indicating the heredity of the model "strong", "weak" or "no"}

\item{model_selection_type}{string indicating the model selection type "JN_org", "LASSO", "APR" (All Possible Regression) or "StepAICc". We recommend using either JN_org or LASSO as the criterion for model selection.}

\item{rm_valruns}{index of validation runs to remove while fitting the model, useful when using cross validation approaches}

\item{y}{the response values for the experiment}
}
\value{
a list of
 \itemize{
       \item ME- active main effects from DSD model selection
       \item 2FI- active 2FI from the DSD model selection
       \item QE- active QE from the DSD model selection
       }
}
\description{
selects the active first-order and second-order terms for the stage 1 design using JN model selection method
}
