% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_lambda_telescoping.R
\name{E_W}
\alias{E_W}
\title{function for computing the moment matrix for a given lambda using Monte Carlo sampling}
\usage{
E_W(
  low,
  high,
  shape1 = 1,
  shape2 = 1,
  num = 1e+05,
  C_dtype,
  model_order,
  me_index = c(),
  qe_index = c(),
  two_fi_index = c()
)
}
\arguments{
\item{low}{a vector of lower bound on the experimental region to compute the moment matrix}

\item{high}{a vector of upper bound on the experimental region to compute the moment matrix}

\item{shape1}{shape of the beta distribution (set at a default of 1)}

\item{shape2}{shape of the beta distribution (set at a default of 1)}

\item{num}{the number of samples in the Monte Carlo estimate}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{model_order}{a string indicating the order of the regression model. "linear","quadratic" (Main Effects+Quadratic Effects), "2FI" (Main Effects + TwoFactor Interactions),"full", "non_standard"- for a user defined model order order}

\item{me_index}{vector of index of Main Effects to be included in the design for model_order "non_standard"}

\item{qe_index}{vector index of Quadratic Effects to be included in the design for model_order "non_standard"}

\item{two_fi_index}{list of index of Two Factor Interactions to be included in the design for model_order "non_standard"}
}
\value{
returns a moment matrix matrix W_mat
 @examples
  E_W(low, high, shape1,shape2,num=100000,C_dtype,model_order,me_index=c(),qe_index=c(),two_fi_index=c())
}
\description{
function for computing the moment matrix for a given lambda using Monte Carlo sampling. The specific Lambda used by default is uniform (tophat prior) weights in a subregion of the experimental region and 0 elsewhere.
}
