% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f_x_cpp}
\alias{f_x_cpp}
\title{function to create a row of the design matrix of a specified model order}
\usage{
f_x_cpp(
  x,
  model_order,
  C_dtype,
  me_index = as.integer(c()),
  qe_index = as.integer(c()),
  two_fi_index_ = NULL
)
}
\arguments{
\item{x}{a vector or row matrix of the settings to generate the design matrix for}

\item{model_order}{a string indicating the order of the regression model. "linear","quadratic" (Main Effects+Quadratic Effects), "2FI" (Main Effects + TwoFactor Interactions),"full", "non_standard"- for a user defined model order order}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{me_index}{vector of index of Main Effects to be included in the design for model_order "non_standard"}

\item{qe_index}{vector index of Quadratic Effects to be included in the design for model_order "non_standard"}

\item{two_fi_index_}{list of index of Two Factor Interactions to be included in the design for model_order "non_standard"}
}
\value{
a row of the design matrix for the specified factor levels and  model_order
 @examples
  f_x(x,model_order="full",C_dtype, me_index=c(),qe_index=c(),two_fi_index=c(),cubic_index=c(),quatric_index=c())
}
\description{
function to create a row of the design matrix of a specified model order
}
