% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sumulation_Functions.R
\name{factornames_standardize}
\alias{factornames_standardize}
\title{function to transform the factor names to a standard notation and backtransform}
\usage{
factornames_standardize(
  design,
  backtransform = FALSE,
  colnames_vector = c(),
  factor_levels = FALSE
)
}
\arguments{
\item{design}{a design matrix}

\item{backtransform}{a boolean value, when TRUE transforms the column names from standard notation "X1"..."Xn" to factor names}

\item{colnames_vector}{a vector mapping the factor names to a standard notation "X1"..."Xn". To be provided when backtransform=TRUE}

\item{factor_levels}{a boolean value, when TRUE indicates that the input matrix is the settings matrix and when FALSE indicates that the input matrix is the design matrix. Design matrix is assumed to contain intercept as the first column.}
}
\value{
a list of:
     \itemize{
       \item colnames_vector- a vector mapping the factor names to a standard notation "X1"..."Xn"
       \item design- the matrix with coded column names
     }
}
\description{
function to transform the factor names to a standard notation "X1"..."Xn" and backtransform
}
