% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_lambda_telescoping.R
\name{fx}
\alias{fx}
\title{function to compute the x'b value for any given x and b}
\usage{
fx(
  x,
  b,
  model_order,
  C_dtype,
  me_index = c(),
  qe_index,
  two_fi_index = c(),
  cubic_index = c(),
  quatric_index = c()
)
}
\arguments{
\item{x}{vector of factor levels in a row of the design}

\item{b}{vector of regression parameters}

\item{model_order}{a string indicating the order of the regression model. "linear","quadratic" (Main Effects+Quadratic Effects), "2FI" (Main Effects + TwoFactor Interactions),"full", "non_standard"- for a user defined model order order}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{me_index}{vector of index of Main Effects to be included in the design for model_order "non_standard"}

\item{qe_index}{vector index of Quadratic Effects to be included in the design for model_order "non_standard"}

\item{two_fi_index}{list of index of Two Factor Interactions to be included in the design for model_order "non_standard"}

\item{cubic_index}{vector of index of Cubic Effects to be included in the design for model_order "non_standard"}

\item{quatric_index}{vector of index of Quatric Effects to be included in the design for model_order "non_standard"}
}
\value{
returns a numeric value numeric value of t(x)%*%b
}
\description{
function to compute the x'b value for any given x and b
}
