% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_X_matrix_v2.0.R
\name{gen_x}
\alias{gen_x}
\title{function to create a starting design of run size n and specified model order}
\usage{
gen_x(
  n,
  C,
  model_order,
  C_dtype,
  type = "non_augment",
  me_index = c(),
  qe_index = c(),
  two_fi_index = c(),
  cubic_index = c(),
  quatric_index = c(),
  design = "Non-Bayesian"
)
}
\arguments{
\item{n}{run size of the design}

\item{C}{List of candidate space for each experimental factor}

\item{model_order}{a string indicating the order of the regression model. "linear","quadratic" (Main Effects+Quadratic Effects), "2FI" (Main Effects + TwoFactor Interactions),"full", "non_standard"- for a user defined model order order}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{type}{a string indicating the type of design "non_augment" or "augment" (augment an existing deisgn)}

\item{me_index}{vector of index of Main Effects to be included in the design for model_order "non_standard"}

\item{qe_index}{vector index of Quadratic Effects to be included in the design for model_order "non_standard"}

\item{two_fi_index}{list of index of Two Factor Interactions to be included in the design for model_order "non_standard"}

\item{cubic_index}{vector of index of Cubic Effects to be included in the design for model_order "non_standard" (although a cubic design can be sampled, current design algorithms don't support cubic models)}

\item{quatric_index}{vector of index of Quatric Effects to be included in the design for model_order "non_standard" (although a quatric design can be sampled, current design algorithms don't support cubic models)}

\item{design}{a string indicating the type of design constructed. "Non-Bayesian" or "Bayesian" optimal design. The major difference is in how signularity of the matrix is evaluated. In a Bayesian optimal design X'X+K(prior variance matrix) is checked for singularity}
}
\value{
a list of factor levels matrix and design matrix of the specified model order
 \itemize{
       \item x- a matrix of factor levels in the design
       \item x_matrix- a design matrix of the specified order including the intercept
       }
 @examples
  gen_x(n,C,model_order="full",C_dtype,type="non_augment",me_index=c(),qe_index=c(),two_fi_index=c(),cubic_index=c(),quatric_index=c(),design="Non-Bayesian")
}
\description{
function to create a starting design of run size n and specified model order
}
