% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_lambda_telescoping.R
\name{individual_bootstrap}
\alias{individual_bootstrap}
\title{function to compute the optimal x* for a given regression function}
\usage{
individual_bootstrap(
  i,
  b,
  fx,
  model_order = "non_standard",
  C_dtype,
  me_index,
  qe_index,
  two_fi_index,
  lower,
  upper,
  nfactor,
  C,
  minimize = TRUE
)
}
\arguments{
\item{i}{is the bootstrap sample index}

\item{b}{vector that is a sample from the multivariate normal distribution of the regression parameters}

\item{fx}{function to compute regression prediction t(x)%*%b}

\item{model_order}{a string indicating the order of the regression model. "linear","quadratic" (Main Effects+Quadratic Effects), "2FI" (Main Effects + TwoFactor Interactions),"full", "non_standard"- for a user defined model order order}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{me_index}{vector of index of Main Effects to be included in the design for model_order "non_standard"}

\item{qe_index}{vector index of Quadratic Effects to be included in the design for model_order "non_standard"}

\item{two_fi_index}{list of index of Two Factor Interactions to be included in the design for model_order "non_standard"}

\item{lower}{a vector of lower limits of the experimental region for all factors (lower level constraint to the optimization)}

\item{upper}{a vector of upper limits of the experimental region for all factors (upper level constraint to the optimization)}

\item{nfactor}{number of experimental factors}

\item{C}{list of candidate space for each experimental factor in the design}

\item{minimize}{a boolean value indicating whether to minimize or maximize. Default value is minimize}
}
\value{
returns a vector of optimum x values for the constrained optimization within the experimental region
}
\description{
function to compute the optimal x* for a given regression function
}
