% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_D_Optimality_telescoping.R
\name{individual_starts}
\alias{individual_starts}
\title{function for initiating the individual coordinate exchange random starts}
\usage{
individual_starts(
  i,
  x_matrix_start,
  x_start,
  n,
  C_dtype,
  C,
  model_order,
  type,
  freeze_rows,
  me_index,
  qe_index,
  two_fi_index,
  cubic_index,
  quatric_index,
  K,
  design,
  use_cpp = TRUE
)
}
\arguments{
\item{i}{index of random starts}

\item{x_matrix_start}{the design matrix of the specified order including the intercept of the random starting}

\item{x_start}{the settings matrix of the random starting design}

\item{n}{total number of runs in the design}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{C}{list of candidate space for each experimental factor in the design}

\item{model_order}{a string indicating the order of the regression model. "linear","quadratic" (Main Effects+Quadratic Effects), "2FI" (Main Effects + TwoFactor Interactions),"full", "non_standard"- for a user defined model order order}

\item{type}{a string indicating the type of design augmentation. "non_augment" or "augment" (augment an existing design)}

\item{freeze_rows}{the rows of the matrix to freeze while augmenting, 0 when type is "non_augment"}

\item{me_index}{vector of index of Main Effects to be included in the design for model_order "non_standard"}

\item{qe_index}{vector index of Quadratic Effects to be included in the design for model_order "non_standard"}

\item{two_fi_index}{list of index of Two Factor Interactions to be included in the design for model_order "non_standard"}

\item{cubic_index}{vector of index of Cubic Effects to be included in the design for model_order "non_standard"}

\item{quatric_index}{vector of index of Quatric Effects to be included in the design for model_order "non_standard"}

\item{K}{a diagonal matrix specifying the prior variance for Bayesian D-optimal design. By default set to 0.001 for potential terms and 0 for primary terms.}

\item{design}{a string indicating the type of design constructed. "Non-Bayesian" or "Bayesian" optimal design. The major difference is in how signularity of the matrix is evaluated. In a Bayesian optimal design X'X+K(prior variance matrix) is checked for singularity}

\item{use_cpp}{a boolean value indicating whether to use the CPP version of coordinate exchange}
}
\value{
a list of
 \itemize{
       \item x_matrix- a Bayesian D-optimal design matrix of the specified order
       \item optimal_value- Optimal value of the objective value of the Bayesian D-optimal design
       \item x- a matrix with factor levels of the Bayesian D-optimal design
       }
 @examples
 \dontrun{
   individual_starts(i,x_matrix_start,x_start,n,C_dtype, C,model_order, type= "augment", freeze_rows=0,me_index = c(),qe_index = c(),two_fi_index = c(),cubic_index = c(),quatric_index = c(),n_starts=500,K,design = "Bayesian")
   }
}
\description{
function for initiating the individual coordinate exchange random starts.
}
