% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_lambda_telescoping.R
\name{individual_starts_compound}
\alias{individual_starts_compound}
\title{function for initiating multiple starts of Adaptive-RSO designs for each random start}
\usage{
individual_starts_compound(
  i,
  x_matrix_start,
  x_start,
  W_mat,
  n,
  C_dtype,
  C,
  model_order,
  type,
  freeze_rows,
  D_opt_design,
  D_opt_x,
  I_opt_design,
  me_index,
  qe_index,
  two_fi_index,
  w,
  D_opt_type,
  K,
  me_index_daug,
  qe_index_daug,
  two_fi_index_daug,
  use_cpp = TRUE
)
}
\arguments{
\item{i}{index of the random start}

\item{x_matrix_start}{the design matrix of the specified order including the intercept of the random starting}

\item{x_start}{the settings matrix of the random starting design}

\item{W_mat}{Moment matrix computed for a given lambda}

\item{n}{total number of runs in the design}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{C}{list of candidate space for each experimental factor in the design}

\item{model_order}{a string indicating the order of the regression model. "linear","quadratic" (Main Effects+Quadratic Effects), "2FI" (Main Effects + TwoFactor Interactions),"full", "non_standard"- for a user defined model order order}

\item{type}{a string indicating the type of design augmentation. "non_augment" or "augment" (augment an existing design)}

\item{freeze_rows}{the rows of the matrix to freeze while augmenting, 0 when type is "non_augment"}

\item{D_opt_design}{the D-optimal design matrix}

\item{D_opt_x}{the settings matrix of the D-optimal design}

\item{I_opt_design}{the I-optimal design matrix}

\item{me_index}{vector of index of Main Effects to be included in the design for model_order "non_standard"}

\item{qe_index}{vector index of Quadratic Effects to be included in the design for model_order "non_standard"}

\item{two_fi_index}{list of index of Two Factor Interactions to be included in the design for model_order "non_standard"}

\item{D_opt_type}{Type of D-optimal design Bayesian or regular D-optimal design. "Bayesian" or "non_Bayesian"}

\item{K}{a diagonal matrix specifying the prior variance in Bayesian D-optimal design set to 0.001 for potential terms and 0 for primary terms}

\item{me_index_daug}{vector of index of Main Effects to be included in the Bayes-D-optimal design for model_order "non_standard"}

\item{qe_index_daug}{vector index of Quadratic Effects to be included in the Bayes-D-optimal design for model_order "non_standard"}

\item{two_fi_index_daug}{list of index of Two Factor Interactions to be included in the Bayes-D-optimal design for model_order "non_standard"}

\item{use_cpp}{a boolean value indicating whether to use the CPP version of coordinate exchange}
}
\value{
returns a list of compound optimal design and its optimal value for a given w and random start
  @examples
  individual_starts_compound(i,x_matrix_start,x_start,W_mat,n,C_dtype,C,model_order, type,freeze_rows,D_opt_design,D_opt_x,I_opt_design,me_index,qe_index,two_fi_index,w,D_opt_type,K,me_index_daug, qe_index_daug,two_fi_index_daug,use_cpp=TRUE)
}
\description{
function for initiating multiple starts of the Adaptive-RSO designs using coordinate exchange algorithm for each random start
}
