% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_unif.R
\name{individual_starts_compound_unif}
\alias{individual_starts_compound_unif}
\title{function for initiating individual starts of the compound uniform design using coordinate exchange.This is the parallel version and works only on linux machines.}
\usage{
individual_starts_compound_unif(
  i,
  x_matrix_start,
  x_start,
  n,
  C_dtype,
  C,
  model_order,
  type,
  freeze_rows,
  D_opt_design,
  D_opt_x,
  x_unifopt,
  unif_crit,
  w,
  use_cpp = FALSE,
  K,
  D_opt_type = "Bayesian",
  aug_type = "not_foldover",
  me_index_daug,
  qe_index_daug,
  two_fi_index_daug,
  update_all_rows = FALSE
)
}
\arguments{
\item{i}{index of the random start}

\item{x_matrix_start}{the design matrix of the specified order including the intercept of the random starting}

\item{x_start}{the settings matrix of the random starting design}

\item{n}{total number of runs in the design (including the augmentations)}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{C}{list of candidate space for each experimental factor in the design}

\item{model_order}{a string indicating the order of the regression model. "linear","quadratic" (Main Effects+Quadratic Effects), "2FI" (Main Effects + TwoFactor Interactions),"full", "non_standard"- for a user defined model order order}

\item{type}{a string indicating the type of design augmentation. "non_augment" or "augment" (augment an existing design)}

\item{freeze_rows}{numeric value indicating the rows to not update}

\item{D_opt_design}{the D-optimal design matrix}

\item{D_opt_x}{the settings matrix of the D-optimal design}

\item{x_unifopt}{the settings matrix of the Uniform-optimal design}

\item{unif_crit}{a string indicating the discrepancy measure for uniform criterion. "MD2" or "WD2"}

\item{w}{a numeric value of weight used to compute the compute the compound optimal design}

\item{use_cpp}{a boolean value indicating whether to use the CPP version of coordinate exchange}

\item{K}{a diagonal matrix specifying the prior variance in Bayesian D-optimal design set to 0.001 for potential terms and 0 for primary terms}

\item{D_opt_type}{a string indicating whether the D-optimal design type is "Bayesian" or "Non_Bayesian"}

\item{aug_type}{a string indicating whether the design has constraint or not. "foldover" or "not_foldover"}

\item{me_index_daug}{vector of index of Main Effects to be included in the D-optimal design for model_order "non_standard"}

\item{qe_index_daug}{vector index of Quadratic Effects to be included in the D-optimal design for model_order "non_standard"}

\item{two_fi_index_daug}{list of index of Two Factor Interactions to be included in the D-optimal design for model_order "non_standard"}

\item{update_all_rows}{a boolean value indicating whether all the rows should be updated or only the augmented rows should be updated}
}
\value{
a list of
 \itemize{
       \item Compound_unif_design- Comound Uniform optimal design matrix of the specified order including the intercept
       \item Compound_unif_obj- The value of the Comound Uniform Optimal objective value
       \item Compound_unif_x- a matrix with factor levels of the Comound Uniform optimal design
       \item Compound_Deff- Bayesian D effciency of the compound uniform design
       \item Compound_Ueff- Uniform Effciency of the compound uniform design
       }
 @examples
 \dontrun{
  individual_starts_compound_unif(i,x_matrix_start,x_start,n,C_dtype,C,model_order, type,freeze_rows,D_opt_design,D_opt_x,x_unifopt,unif_crit,w,use_cpp=FALSE,K,D_opt_type="Bayesian",aug_type="not_foldover",me_index_daug, qe_index_daug,two_fi_index_daug,update_all_rows=FALSE)
   }
 @export
}
\description{
function for initiating individual starts of the compound uniform design using coordinate exchange.This is the parallel version and works only on linux machines.
}
