% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_lambda_telescoping.R
\name{individual_starts_i_opt}
\alias{individual_starts_i_opt}
\title{function for initiating individual coordinate exchange for I optimal design for the parallel code.}
\usage{
individual_starts_i_opt(
  i,
  x_matrix_start,
  x_start,
  W_mat,
  n,
  C_dtype,
  C,
  model_order,
  type,
  me_index,
  qe_index,
  two_fi_index,
  use_cpp = TRUE
)
}
\arguments{
\item{i}{random start index}

\item{x_matrix_start}{the design matrix of the specified order including the intercept of the random starting}

\item{x_start}{the settings matrix of the random starting design}

\item{W_mat}{Moment matrix computed for a given lambda}

\item{n}{total number of runs in the design}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{C}{list of candidate space for each experimental factor in the design}

\item{model_order}{a string indicating the order of the regression model. "linear","quadratic" (Main Effects+Quadratic Effects), "2FI" (Main Effects + TwoFactor Interactions),"full", "non_standard"- for a user defined model order order}

\item{type}{a string indicating the type of design augmentation. "non_augment" or "augment" (augment an existing design)}

\item{me_index}{vector of index of Main Effects to be included in the design for model_order "non_standard"}

\item{qe_index}{vector index of Quadratic Effects to be included in the design for model_order "non_standard"}

\item{two_fi_index}{list of index of Two Factor Interactions to be included in the design for model_order "non_standard"}

\item{use_cpp}{a boolean value indicating whether to use the CPP version of coordinate exchange}
}
\value{
a list of
 \itemize{
       \item x_matrix- a I-optimal design matrix of the specified order
       \item optimal_value- Optimal value of the objective value of the I-optimal design
       \item x- a matrix with factor levels of the I-optimal design
       }
 @examples
  individual_starts_i_opt=function(i,x_matrix_start,x_start,W_mat,n,C_dtype, C,model_order,type,me_index,qe_index,two_fi_index)
}
\description{
function for initiating individual coordinate exchange for I optimal design for the parallel code.
}
