% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JN_method.R
\name{JN_model_selection}
\alias{JN_model_selection}
\title{a function for Jones and Nachtsheim (JN) model selection method}
\usage{
JN_model_selection(
  dsd_matrix,
  y,
  alpha = 0.2,
  heredity = "strong",
  fake_factor_index,
  C_dtype,
  model_selection_type = "JN_org"
)
}
\arguments{
\item{dsd_matrix}{a DSD matrix for model selection}

\item{y}{a vector of response}

\item{alpha}{a numeric value the signficance level for model selection, set at the value recommended in the JN 2017 paper}

\item{heredity}{a string indicating the heredity for model selection. One of "Strong", "weak", and "no". Heredity determines the number of second order terms considered in model selection.}

\item{fake_factor_index}{a vector indeicating the index of facle factor terms in the design matrix}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{model_selection_type}{a string indicating the type of model selection method adopted for second order terms. One of "JN_org"- method proposed in the paper, "LASSO"- LASSO with AICc, "APR"- all possible regression, not recommended for large designs and "StepAICc"- stepwise AICc}
}
\value{
a list of:
   \itemize{
     \item ME Index- index of active Main Effects
     \item Beta_me- parameter estimates of the main effects
     \item Tstat- t-statistic of the hypothesis test of regression parameters
     \item p-value- p-value of the significance test of regression parameters
     \item active_se- regression matrix of active second order terms
     \item beta_se- parameter estimates of the second order terms
   }
}
\description{
a function for Jones and Nachtsheim model selection method. Works only for foldover designs.
}
