% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_X_matrix_v2.0.R
\name{phi_x}
\alias{phi_x}
\title{an internal function that computes the intermediate steps in I-optimal design}
\usage{
phi_x(x_i, x_j, D, W)
}
\arguments{
\item{x_i}{a row of the design matrix}

\item{x_j}{a row of the design matrix}

\item{D}{current X'X inverse matrix}

\item{W}{Moment matrix}
}
\value{
a matrix
}
\description{
an internal function that computes the intermediate steps in computing the change in determinant value for each coordinate exchange in the construction of I-optimal designs
}
