% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_X_matrix_v2.0.R
\name{pick_rand}
\alias{pick_rand}
\title{function to sample from the candidate space}
\usage{
pick_rand(C)
}
\arguments{
\item{C}{List of candidate space for each experimental factor}
}
\value{
A row of the factor levels matrix sampled from the candidate space
}
\description{
function to sample from the candidate space
}
