% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank2inverse.R
\name{rank2_inverse}
\alias{rank2_inverse}
\title{function for rank 2 inverse to update the inverse of X'X matrix for each coordinate exchange iteration}
\usage{
rank2_inverse(x, x_i, D)
}
\arguments{
\item{x}{the row vector with the new coordinate value}

\item{x_i}{the current row vector that needs to be exchanged}

\item{D}{Current (X'X)inv matrix}
}
\value{
A matrix with the updated (X'X)inv if the exhange leads to a non singular X else a character string "Singular"
 @examples
  rank2_inverse(x,x_i,D)
}
\description{
function for rank 2 inverse to update the inverse of X'X matrix for each coordinate exchange iteration
}
