% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sumulation_Functions.R
\name{stage2_aug}
\alias{stage2_aug}
\title{Augments the starting design DSD with compound uniform design for a specified number of runs}
\usage{
stage2_aug(
  active_me,
  active_2FI,
  active_qe,
  design_matrix,
  fake_factor_index,
  naug,
  se_names,
  C_dtype,
  C,
  heredity = "no",
  n_starts = 100,
  n_starts_compound = 10,
  design_aug_type = "compound_unif",
  unif_crit = "MD2",
  U_opt_tresh = 0.8,
  D_opt_tresh = 0.8,
  w_tresh = 0.02,
  use_w = FALSE,
  w_candidate = 0,
  rng = c(),
  num_cores = NA
)
}
\arguments{
\item{active_me}{vector of active ME from the stage 1 regression}

\item{active_2FI}{vector of active 2FI from the stage 1 regression}

\item{active_qe}{vector of active QE from the stage 1 regression}

\item{design_matrix}{DSD matrix with fake factors}

\item{fake_factor_index}{index of fake factor columns}

\item{naug}{number of runs to augment the starting design by}

\item{se_names}{column names of the design matrix}

\item{C_dtype}{data type of the factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{C}{list of candidate space for each experimental factor in the design}

\item{heredity}{string indicating the heredity of the model "strong", "weak" or "no". Determines whether inactive factors should be dropped.}

\item{n_starts}{number of random starts when constructing D-optimal designs}

\item{n_starts_compound}{number of random starts when constructing Compound Uniform (robust) designs}

\item{design_aug_type}{if "compound_unif" computes the compound uniform criterion}

\item{unif_crit}{Uniform discrepancy criterion either "MD2" (default) and "WD2"}

\item{U_opt_tresh}{a numeric value providing the lower bound of the threshold in bisection search for the Uniform optimal criteria. Default value is 0.8}

\item{D_opt_tresh}{a numeric value providing the lower bound of the threshold in bisection search for the D-optimal criteria. Default value is 0.8}

\item{w_tresh}{the threshold for difference in weights w for the bisection search to converge. The default value is 0.02}

\item{use_w}{a boolean value indicating whether w_candidate should be used or the bisection search method for finding the compound optimal designs}

\item{w_candidate}{if use_w is TRUE, a vector of weights w to compute the Uniform optimal design for}

\item{rng}{random seed for random number generation}

\item{num_cores}{number of cores for parallel processing. Default value is NA, in which case Sys.getenv("PBS_RESC_TOTAL_PROCS") is used to get the number of cores in a HPC. Otherwise, the user needs to provide a value. Use 1 for serial execution.}
}
\value{
A list of:
    \itemize{
      \item duag_design-D-optimal desugn for the augmented runs
      \item duag_x- factor levels matrix of the D-optimal design for the augmented runs
      \item compound_unif_design- A list of compound uniform optimal designs for each weigth w
      \item design_eff_results- a matrix siummarizing the design effciencies of the compound optimal design
    }
}
\description{
Augments the starting design DSD with compound uniform design for a specified number of runs
}
