% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sumulation_Functions.R
\name{stage2_aug_analysis}
\alias{stage2_aug_analysis}
\title{fits a GAUSS-LASSO regression model (LASSO with AICc is used for model selection and an OLS model is fit for the identified terms), on the stage 2 design and returns the lack of fit test result}
\usage{
stage2_aug_analysis(design_matrix, y, se_names, lof_alpha = 0.05)
}
\arguments{
\item{design_matrix}{design matrix from Stage 2 (the matrix with all the regression terms)}

\item{y}{a vector of responses for the experiment}

\item{se_names}{column names of the design matrix}

\item{lof_alpha}{significance level for the Lack of Fit test}
}
\value{
a list of:
 \itemize{
    \item beta- active terms from model and the Gauss-LASSO estimates for the active terms
    \item Sigma- the variance-covariance matrix for the regression model
    \item F_stat_lof- F statistic of lack of fit test
    \item F_crit_lof- F critical value of lack of fit test
    \item lof_result- lack of fit test result
    \item resid- the residuals of the Gauss LASSO model
   }
}
\description{
fits a GAUSS-LASSO regression model (LASSO with AICc is used for model selection and an OLS model is fit for the identified terms), on the stage 2 design and returns the lack of fit test result
}
