% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sumulation_Functions.R
\name{stage3_analysis}
\alias{stage3_analysis}
\title{fits a GAUSS LASSO regression model with AICc for model selection to the stage 3 design and returns the regression parameters and optimal response.}
\usage{
stage3_analysis(
  design_matrix,
  design_x,
  y,
  se_names,
  C,
  C_dtype,
  minimize = TRUE
)
}
\arguments{
\item{design_matrix}{stage 3 design matrix of a given order}

\item{design_x}{stage 3 settings matrix}

\item{y}{a vector of response for the experiments}

\item{se_names}{column names of the stage 3 design}

\item{C}{list of of candidate points for each factor in the experiment}

\item{C_dtype}{data type of the factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{minimize}{a boolean value indicating whether to minimize or maximize. Default value is minimize}
}
\value{
\itemize{
    \item beta- the regression parameters from the Gauss Lasso model
    \item x_star- the predicted optimal location for the response surface
    \item y_min- the predicted optimal response for the response surface
  }
}
\description{
fits a GAUSS LASSO regression model with AICc for model selection to the stage 3 design and returns the regression parameters and optimal response.
}
