% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sumulation_Functions.R
\name{stage3_aug}
\alias{stage3_aug}
\title{constructs the Adaptive-RSO designs in Stage 3 design}
\usage{
stage3_aug(
  beta,
  design_matrix,
  design_x,
  Sigma,
  se_names,
  naug,
  n_starts = 100,
  n_starts_compound = 100,
  B = 5000,
  C_dtype,
  C,
  me_index_daug,
  qe_index_daug,
  two_fi_index_daug,
  use_cpp = TRUE,
  D_opt_tresh = 0.8,
  I_opt_tresh = 0.8,
  w_tresh = 0.02,
  minimize = TRUE,
  rng = c(),
  num_cores = NA
)
}
\arguments{
\item{beta}{a vector of regression parameter estimates from stage 2 analysis}

\item{design_matrix}{design matrix from Stage 2}

\item{design_x}{settings matrix from Stage 2}

\item{Sigma}{variance covariance matrix of the regression model from stage 2}

\item{se_names}{design matrix column names from stage 2}

\item{naug}{number of runs to augment design by in stage 3}

\item{n_starts}{number of random starts for constructing D-optimal and I-optimal designs}

\item{n_starts_compound}{number of random starts for design constructing compound optimal designs}

\item{B}{number of bootstrap samples for x* confidence region}

\item{C_dtype}{data type of the factors in the design either "cont"(continuous) or "cat"(categorical)}

\item{C}{list of of candidate points for each factor in the experiment}

\item{me_index_daug}{vector of main effects index in the D-optimal design}

\item{qe_index_daug}{vector of quadratic index in the D-optimal design}

\item{two_fi_index_daug}{list of two factor interaction index in the D-optimal design}

\item{use_cpp}{boolean indicating whether to use CPP code or not}

\item{D_opt_tresh}{a numeric value providing the lower bound of the threshold in bisection search for the D-optimal criteria. Default value is 0.8}

\item{I_opt_tresh}{a numeric value providing the lower bound of the threshold in bisection search for the I-optimal criteria. Default value is 0.8}

\item{w_tresh}{the threshold for difference in weights w for the bisection search to converge. The default value is 0.02}

\item{minimize}{a boolean value indicating whether to minimize or maximize. Default value is minimize}

\item{rng}{seed for random number generation for the random design construction starts}

\item{num_cores}{number of cores for parallel processing. Default value is NA, in which case Sys.getenv("PBS_RESC_TOTAL_PROCS") is used to get the number of cores in a HPC. Otherwise, the user needs to provide a value. Use 1 for serial execution.}
}
\value{
returns a list of:
      \itemize{
         \item D_aug- list of Bayesian D-optimal augmentation design matrix and factor levels matrix of the stage 2 design for the specified number of augmentation runs
         \item Boostrap_results- matrix containing the subregion of the design space idetified for exploitation
         \item I_aug- list of I-optimal augmentation design matrix and factor levels matrix of the stage 2 design for the specified number of augmentation runs
         \item compound_weff- list of the best exploration exploitation designs for a set of weights w evaluated by bisection search
         \item K- prior variance matrix for Bayesian D-optimal design
         \item W_mat- moment matrix for I-lambda designs
         \item x_star_beta- optimal location x* for the current estimate beta
         }
}
\description{
constructs the Adaptive-RSO designs in Stage 3 design
}
