% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JN_method.R
\name{stepAICc}
\alias{stepAICc}
\title{function for stepwise AICc model selection}
\usage{
stepAICc(y, XME, XSE, maxterms)
}
\arguments{
\item{y}{a vector of response}

\item{XME}{a matrix of active Main Effects from stage 1 JN method}

\item{XSE}{a mtraix of second order terms}

\item{maxterms}{a numeric value indicating the max terms in the design}
}
\value{
a list of:
     \itemize{
       \item x_se- the matrix of second order terms for the model chosen from stepwise AICc
       \item bestSSE- the SSE of the chosen model
       \item minAICc- AICCc of the chosen model
       \item best_beta- the parameter estimates of the chosen model
     }
}
\description{
function for stepwise AICc model selection on the second order terms in the JN model selection method
}
