% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sumulation_Functions.R
\name{transform_designspace}
\alias{transform_designspace}
\title{function to transform the design space}
\usage{
transform_designspace(design_x, std = TRUE, lower, upper)
}
\arguments{
\item{design_x}{a settings matrix}

\item{std}{a boolean value, when TRUE transforms design from (a,b) to (-1,1) and when FALSE transforms (-1,1) to (a,b)}

\item{lower}{a vector of lower levels for the factors in the experiment}

\item{upper}{a vector of upper levels for the factors in the experiment}
}
\value{
a transformed matrix of design factor levels
}
\description{
a function that transforms the design space from (a,b) to (-1,1) or from (-1,1) to (a,b)
}
