% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sumulation_Functions.R
\name{update_terms_stage3}
\alias{update_terms_stage3}
\title{drops inactive factors from the design for a given heredity}
\usage{
update_terms_stage3(
  drop_inactive_factors,
  heredity,
  design_matrix,
  design_x,
  beta,
  se_names,
  candidate_factors,
  C,
  C_dtype
)
}
\arguments{
\item{drop_inactive_factors}{boolean value indicating whether to drop inactive factors}

\item{heredity}{string indicating the heredity of the model "strong", "weak" or "no". Determines which inactive effect should be dropped.}

\item{design_matrix}{design matrix from stage 2 augmentation}

\item{design_x}{settings matrix from stage 2 augmentation}

\item{beta}{the regression parameters from stage 2 analysis}

\item{se_names}{vector of column names of the design matrix}

\item{candidate_factors}{vector of names of experimental factors}

\item{C}{list of candidate space for each experimental factor in the design}

\item{C_dtype}{a string vector indicating the data type of the experimental factors in the design either "cont"(continuous) or "cat"(categorical)}
}
\value{
a list of
      \itemize{
         \item design_matrix- updated design matrix with the specified heredity
         \item design_x- updated factor level matrix
         \item me_index- updated main effects index vector for the design matrix
         \item qe_index- updated quadratic effects index vector for the design matrix
         \item two_fi_index- updated list of two factor interaction index
         \item C- updated candidate points list
         \item C_dtype- updated vector of datatypes of the experimental factors
         \item se_names- updated vector of column names of the design matrix
      }
}
\description{
drops inactive factors from the design for a given heredity
}
