% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{v_x_cpp}
\alias{v_x_cpp}
\title{an internal function that computes the intermediate steps in D-optimal designs}
\usage{
v_x_cpp(x_i, x_j, D)
}
\arguments{
\item{x_i}{a row of the design matrix}

\item{x_j}{a row of the design matrix}

\item{D}{current X'X inverse matrix}
}
\value{
a matrix
}
\description{
an internal function that computes the intermediate steps in computing the change in determinant value for each coordinate exchange in the construction of D-optimal designs
}
